refactor(api-reference): update import paths and enhance test structure
