+++ packages/api-reference/src/components/Content/Introduction/Introduction.test.ts
@@ -100,7 +100,7 @@ describe('Introduction', () => {
     })
 
     const downloadLink = wrapper.findComponent(DownloadLink)
-    expect(downloadLink.props('specTitle')).toBe('hello-world-api')
+    expect(downloadLink.props('filename')).toBe('hello-world-api')
   })
 
   it('shows version badge when version exists', () => {
