+++ packages/api-reference/src/components/Section/CompactSection.vue
@@ -2,10 +2,10 @@
 import { ScalarIcon } from '@scalar/components'
 import { nextTick, ref, useId, watch } from 'vue'
 
-import { scrollToId } from '../../helpers'
-import { useNavState } from '../../hooks'
-import Anchor from '../Anchor/Anchor.vue'
-import Section from './Section.vue'
+import Anchor from '@/components/Anchor/Anchor.vue'
+import Section from '@/components/Section/Section.vue'
+import { scrollToId } from '@/helpers/scroll-to-id'
+import { useNavState } from '@/hooks/useNavState'
 
 const props = defineProps<{
   id: string
