+++ packages/api-reference/src/features/Search/SearchModal.vue
@@ -12,11 +12,15 @@ import type { FuseResult } from 'fuse.js'
 import { nanoid } from 'nanoid'
 import { ref, toRef, watch } from 'vue'
 
-import { lazyBus } from '../../components/Content/Lazy/lazyBus'
-import SidebarHttpBadge from '../../components/Sidebar/SidebarHttpBadge.vue'
-import { scrollToId } from '../../helpers'
-import { useSidebar } from '../../hooks'
-import { useSearchIndex, type EntryType, type FuseData } from './useSearchIndex'
+import { lazyBus } from '@/components/Content/Lazy/lazyBus'
+import SidebarHttpBadge from '@/components/Sidebar/SidebarHttpBadge.vue'
+import {
+  useSearchIndex,
+  type EntryType,
+  type FuseData,
+} from '@/features/Search/useSearchIndex'
+import { scrollToId } from '@/helpers/scroll-to-id'
+import { useSidebar } from '@/hooks/useSidebar'
 
 const props = defineProps<{
   parsedSpec: Spec
