+++ packages/api-reference/src/helpers/get-variable-names.ts
