--- packages/api-reference/src/helpers/hasModels.ts
@@ -1,15 +1,0 @@
-import type { Spec } from '@scalar/types/legacy'
-
-import { getModels } from './getModels'
-
-export const hasModels = (spec?: Spec) => {
-  if (!spec) {
-    return false
-  }
-
-  if (Object.keys(getModels(spec) ?? {}).length) {
-    return true
-  }
-
-  return false
-}
