--- packages/api-reference/src/helpers/hasSecuritySchemes.ts
@@ -1,10 +1,0 @@
-import type { Spec } from '@scalar/types/legacy'
-
-export const hasSecuritySchemes = (spec?: Spec) => {
-  // TODO: Show security schemes
-  if (Object.keys(spec?.components?.securitySchemes ?? {}).length) {
-    return true
-  }
-
-  return false
-}
