--- packages/api-reference/src/helpers/index.ts
@@ -1,15 +1,0 @@
-export { deepMerge } from './deepMerge'
-export { getHeadingsFromMarkdown } from './getHeadingsFromMarkdown'
-export { getLowestHeadingLevel } from './getLowestHeadingLevel'
-export { getModels } from './getModels'
-export { getVariableNames } from './getVariableNames'
-export { hasModels } from './hasModels'
-export { hasSecuritySchemes } from './hasSecuritySchemes'
-export { hasWebhooks } from './hasWebhooks'
-export { isMacOs } from './isMacOs'
-export { mapFromObject } from './mapFromObject'
-export { mergeAllObjects } from './mergeAllObjects'
-export { parse } from './parse'
-export { scrollToId } from './scrollToId'
-export { sleep } from './sleep'
-export { downloadSpecBus, downloadSpecFile } from './specDownloads'
