+++ packages/api-reference/src/helpers/is-mac-os.ts
