+++ packages/api-reference/src/helpers/merge-all-objects.ts
