+++ packages/api-reference/src/helpers/sleep.ts
@@ -1,7 +1,5 @@
 /**
  * Little helper for sleeping for x milliseconds
  * an async await friendly setTimeout
- *
- * @param ms - time to sleep in ms
  */
 export const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms))
