+++ packages/api-reference/src/hooks/useSidebar.ts
@@ -1,18 +1,13 @@
+import { lazyBus } from '@/components/Content/Lazy/lazyBus'
+import { scrollToId } from '@/helpers/scroll-to-id'
+import { useNavState } from '@/hooks/useNavState'
+import { getHeadingsFromMarkdown, getLowestHeadingLevel } from '@/libs/markdown'
+import { getModels, hasModels, hasWebhooks } from '@/libs/openapi'
 import { isOperationDeprecated } from '@/libs/operation'
 import { ssrState } from '@scalar/oas-utils/helpers'
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { Spec, Tag, TransformedOperation } from '@scalar/types/legacy'
 import { computed, reactive, ref, watch } from 'vue'
-import { lazyBus } from '../components/Content/Lazy/lazyBus'
-import {
-  getHeadingsFromMarkdown,
-  getLowestHeadingLevel,
-  getModels,
-  hasModels,
-  hasWebhooks,
-  scrollToId,
-} from '../helpers'
-import { useNavState } from './useNavState'
 
 export type SidebarEntry = {
   id: string
