+++ packages/api-reference/tsconfig.build.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["dist", "tests", "vite.*.ts", "vitest.*.ts", "**/*.test.ts", "**/*.spec.ts"],
+  "exclude": ["dist", "vite.*.ts", "vitest.*.ts", "**/*.test.ts", "**/*.spec.ts", "test"],
   "compilerOptions": {
     "skipLibCheck": true,
     "declaration": true,
